+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+### 0.0.41 "Eagle" (12.03.2015 - ...)
+* `column` for column names
+* ``JavaScript expression`` for JavaScript
+
+
 ### 0.0.40 "Sapsan" (24.01.2015 - 06.03.2015)
 
 * IF EXISTS() and subqueries